requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-rest-swagger==2.0.7
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2018.5
requests==2.19.1
whitenoise==4.1.2